переписала TableModelListener в контроллер
