+++ summer_practice/src/main/java/project/orange/Inter.java
@@ -281,17 +281,6 @@ public class Inter extends JFrame {
             }
 
             graphMatrix = n.drawMatrix();
-            graphMatrix.getModel().addTableModelListener(new TableModelListener() {
-                @Override
-                public void tableChanged(TableModelEvent e) {
-                    int row = e.getFirstRow();
-                    int column = e.getColumn();
-                    if (row == 0 || column == 0 || row == column) { return; }
-                    TableModel model = (TableModel)e.getSource();
-                    Object data = model.getValueAt(row, column);
-                    n.weightChange(row - 1, column - 1, Integer.valueOf(data.toString()));
-                }
-            });
             startMatrix.add(graphMatrix);
 
             lgraph.setText(input);
@@ -310,17 +299,6 @@ public class Inter extends JFrame {
         }
 
         graphMatrix = n.drawMatrix();
-        graphMatrix.getModel().addTableModelListener(new TableModelListener() {
-            @Override
-            public void tableChanged(TableModelEvent e) {
-                int row = e.getFirstRow();
-                int column = e.getColumn();
-                if (row == 0 || column == 0 || row == column) { return; }
-                TableModel model = (TableModel)e.getSource();
-                Object data = model.getValueAt(row, column);
-                n.weightChange(row - 1, column - 1, Integer.valueOf(data.toString()));
-            }
-        });
         startMatrix.add(graphMatrix);
 
         lgraph.setText(n.getCurrentState());
@@ -406,17 +384,6 @@ public class Inter extends JFrame {
         }
 
         graphMatrix = n.drawMatrix();
-        graphMatrix.getModel().addTableModelListener(new TableModelListener() {
-            @Override
-            public void tableChanged(TableModelEvent e) {
-                int row = e.getFirstRow();
-                int column = e.getColumn();
-                if (row == 0 || column == 0 || row == column) { return; }
-                TableModel model = (TableModel)e.getSource();
-                Object data = model.getValueAt(row, column);
-                n.weightChange(row - 1, column - 1, Integer.valueOf(data.toString()));
-            }
-        });
         startMatrix.add(graphMatrix);
 
         lgraph.setText(input);
@@ -433,17 +400,6 @@ public class Inter extends JFrame {
         }
         startMatrix.remove(graphMatrix);
         graphMatrix = n.drawMatrix();
-        graphMatrix.getModel().addTableModelListener(new TableModelListener() {
-            @Override
-            public void tableChanged(TableModelEvent e) {
-                int row = e.getFirstRow();
-                int column = e.getColumn();
-                if (row == 0 || column == 0 || row == column) { return; }
-                TableModel model = (TableModel)e.getSource();
-                Object data = model.getValueAt(row, column);
-                n.weightChange(row - 1, column - 1, Integer.valueOf(data.toString()));
-            }
-        });
         startMatrix.add(graphMatrix);
         this.revalidate();
 
@@ -458,17 +414,6 @@ public class Inter extends JFrame {
         }
         startMatrix.remove(graphMatrix);
         graphMatrix = n.drawMatrix();
-        graphMatrix.getModel().addTableModelListener(new TableModelListener() {
-            @Override
-            public void tableChanged(TableModelEvent e) {
-                int row = e.getFirstRow();
-                int column = e.getColumn();
-                if (row == 0 || column == 0 || row == column) { return; }
-                TableModel model = (TableModel)e.getSource();
-                Object data = model.getValueAt(row, column);
-                n.weightChange(row - 1, column - 1, Integer.valueOf(data.toString()));
-            }
-        });
         startMatrix.add(graphMatrix);
         this.revalidate();
     }
